Feat: 루틴 상세 조회 기능 추가 #6
