+++ build.gradle
@@ -18,6 +18,9 @@ dependencies {
     implementation 'org.springframework:spring-webmvc:5.3.30' // DispatcherServlet 및 MVC 기능
     implementation 'org.springframework:spring-jdbc:5.3.30' // JDBC 및 트랜잭션 관리
 
+    // JSON 직렬화를 위한 Jackson
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.1'
+
     // 데이터베이스 드라이버 (MySQL 예시)
     implementation 'mysql:mysql-connector-java:8.0.33'
