+++ src/main/java/com/kbulkup/auth/controller/LogoutController.java
@@ -1,0 +1,4 @@
+package com.kbulkup.auth.controller;
+
+public class LogoutController {
+}
