+++ src/main/java/com/kbulkup/auth/dto/SignupRequestDTO.java
@@ -1,0 +1,4 @@
+package com.kbulkup.auth.dto;
+
+public class SignupRequestDTO {
+}
