+++ src/main/java/com/kbulkup/common/config/SecurityConfig.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.config;
+
+public class SecurityConfig {
+}
