+++ src/main/java/com/kbulkup/common/dto/ApiResponse.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.dto;
+
+public class ApiResponse {
+}
