+++ src/main/java/com/kbulkup/common/exception/CommonExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.exception;
+
+public class CommonExceptionHandler {
+}
