+++ src/main/java/com/kbulkup/common/exception/CustomException.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.exception;
+
+public class CustomException {
+}
