+++ src/main/java/com/kbulkup/common/security/CustomUserDetailsService.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.security;
+
+public class CustomUserDetailsService {
+}
