+++ src/main/java/com/kbulkup/common/util/OAuthUtils.java
@@ -1,0 +1,4 @@
+package com.kbulkup.common.util;
+
+public class OAuthUtils {
+}
