+++ src/main/java/com/kbulkup/routine/mapper/TraineeRoutineMapper.java
@@ -1,0 +1,18 @@
+package com.kbulkup.routine.mapper;
+
+import com.kbulkup.routine.dto.TraineeRoutineDetailResponseDTO;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+
+public interface TraineeRoutineMapper {
+
+    /**
+     * 루틴 상세 정보 조회
+     *
+     * @param routineId 조회할 루틴 ID
+     * @return TraineeRoutineDetailResponseDTO
+     */
+    TraineeRoutineDetailResponseDTO findRoutineDetailById(@Param("routineId") Long routineId);
+
+}
