+++ src/main/java/com/kbulkup/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.kbulkup.user.controller;
+
+public class UserController {
+}
