+++ src/main/java/com/kbulkup/user/domain/User.java
@@ -1,0 +1,4 @@
+package com.kbulkup.user.domain;
+
+public class User {
+}
