+++ src/main/java/com/kbulkup/user/dto/UserResponseDTO.java
@@ -1,0 +1,4 @@
+package com.kbulkup.user.dto;
+
+public class UserResponseDTO {
+}
