+++ src/main/java/com/kbulkup/user/mapper/UserMapper.java
@@ -1,0 +1,6 @@
+package com.kbulkup.user.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+
+public interface UserMapper {
+}
