+++ src/main/java/com/kbulkup/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.kbulkup.user.service;
+
+public class UserService {
+}
