--- src/main/java/com/kubulkup/auth/controller/SignupController.java
@@ -1,4 +1,0 @@
-package com.kubulkup.auth.controller;
-
-public class SignupController {
-}
