--- src/main/java/com/kubulkup/auth/dto/LogoutResponseDTO.java
@@ -1,13 +1,0 @@
-package com.kubulkup.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-
-@Builder
-public class LogoutResponseDTO {
-    private String message; // 예: "로그아웃 성공"
-}
