--- src/main/java/com/kubulkup/auth/dto/SignupRequestDTO.java
@@ -1,4 +1,0 @@
-package com.kubulkup.auth.dto;
-
-public class SignupRequestDTO {
-}
