--- src/main/java/com/kubulkup/auth/dto/SignupResponseDTO.java
@@ -1,4 +1,0 @@
-package com.kubulkup.auth.dto;
-
-public class SignupResponseDTO {
-}
