--- src/main/java/com/kubulkup/common/config/SecurityConfig.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.config;
-
-public class SecurityConfig {
-}
