--- src/main/java/com/kubulkup/common/dto/ApiResponse.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.dto;
-
-public class ApiResponse {
-}
