--- src/main/java/com/kubulkup/common/exception/CommonExceptionHandler.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.exception;
-
-public class CommonExceptionHandler {
-}
