--- src/main/java/com/kubulkup/common/exception/CustomException.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.exception;
-
-public class CustomException {
-}
