--- src/main/java/com/kubulkup/common/security/CustomUserDetailsService.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.security;
-
-public class CustomUserDetailsService {
-}
