--- src/main/java/com/kubulkup/common/security/JwtAuthenticationFilter.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.security;
-
-public class JwtAuthenticationFilter {
-}
