--- src/main/java/com/kubulkup/common/security/JwtTokenProvider.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.security;
-
-public class JwtTokenProvider {
-}
