--- src/main/java/com/kubulkup/common/util/OAuthUtils.java
@@ -1,4 +1,0 @@
-package com.kubulkup.common.util;
-
-public class OAuthUtils {
-}
