--- src/main/java/com/kubulkup/user/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.kubulkup.user.controller;
-
-public class UserController {
-}
