--- src/main/java/com/kubulkup/user/domain/User.java
@@ -1,4 +1,0 @@
-package com.kubulkup.user.domain;
-
-public class User {
-}
