--- src/main/java/com/kubulkup/user/dto/UserRequestDTO.java
@@ -1,4 +1,0 @@
-package com.kubulkup.user.dto;
-
-public class UserRequestDTO {
-}
