--- src/main/java/com/kubulkup/user/dto/UserResponseDTO.java
@@ -1,4 +1,0 @@
-package com.kubulkup.user.dto;
-
-public class UserResponseDTO {
-}
