--- src/main/java/com/kubulkup/user/mapper/UserMapper.java
@@ -1,7 +1,0 @@
-package com.kubulkup.user.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface UserMapper {
-}
