--- src/main/java/com/kubulkup/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.kubulkup.user.service;
-
-public class UserService {
-}
