+++ src/main/resources/config/db.properties
@@ -1,4 +1,4 @@
 db.driverClassName=com.mysql.cj.jdbc.Driver
-db.url=jdbc:mysql://localhost:3306/KBulkUP_DB?serverTimezone=UTC&characterEncoding=UTF-8
+db.url=jdbc:mysql://localhost:3306/kbulkupdb?serverTimezone=UTC&characterEncoding=UTF-8
 db.username=kbulkup
 db.password=kbulkup12!@
