+++ src/main/resources/mappers/TraineeRoutineMapper.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.kbulkup.routine.mapper.TraineeRoutineMapper">
+
+    <resultMap id="RoutineDetailMap" type="com.kbulkup.routine.dto.TraineeRoutineDetailResponseDTO">
+    <result property="routineTitle" column="routine_title"/>
+        <result property="routineDescription" column="routine_description"/>
+        <result property="routineScore" column="routine_score"/>
+        <result property="category" column="category"/>
+        <result property="level" column="level"/>
+    </resultMap>
+
+    <select id="findRoutineDetailById" resultMap="RoutineDetailMap" parameterType="map">
+        SELECT
+            r.title AS routine_title,
+            r.description AS routine_description,
+            r.score AS routine_score,
+            t.category,
+            t.level
+        FROM routines r
+                 JOIN trainings t ON r.training_id = t.training_id
+        WHERE r.routine_id = #{routineId}
+    </select>
+
+</mapper>
