+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -9,7 +9,7 @@
 
     <mvc:annotation-driven />
 
-    <context:component-scan base-package="com.kubulkup" />
+    <context:component-scan base-package="com.kbulkup.routine.controller" />
 
     <bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
         <property name="prefix" value="/WEB-INF/views/" /> <property name="suffix" value=".jsp" /> </bean>
