+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -26,7 +26,7 @@
         <constructor-arg ref="sqlSessionFactory" />
     </bean>
 
-    <mybatis-spring:scan base-package="com.kubulkup" />
+    <mybatis-spring:scan base-package="com.kbulkup.routine.mapper" />
 
-    <context:component-scan base-package="com.kubulkup" />
+    <context:component-scan base-package="com.kbulkup" />
 </beans>
